Restructuring of the project
Rewritten a lot of stuff, removed the unnecessary things, and made it a lot easier to work with file formats.
Some stuff isn't here anymore, like the libraries for sound, a lot of functionality in the sprite editor and a lot of custom code for 'asmhacks'. Those will all come back in future commits.